moved user render-prop to fire-fetch
