+++ src/fire-fetch/index.js
@@ -6,4 +6,4 @@ export {
 export { default as RootRef, GetRootRef, withRootRef } from './root-ref';
 export { default as FirebaseRef } from './firebase-ref';
 export { default as FirebaseQuery } from './firebase-query';
-export { default as AuthListener } from './auth-listener';
+export { default as AuthListener, User, withUser } from './auth-listener';
